feat: display on top banner when a team wins
