+++ packages/frontend/src/extra/types.ts
@@ -2,6 +2,9 @@ import { Socket } from "socket.io-client";
 // @ts-ignore
 import { CodenamesApp } from "../components/codenames-app/codenames-app";
 
+// Types
+// ------------------
+
 /**
  * Enumerator for possible cell colors.
  */
@@ -141,3 +144,20 @@ export interface UnfinishedRoom {
   turn: Color;
   currentGameId?: string;
 }
+
+// Constants
+// -----------------
+
+// Constants for how many words of each color should be used
+export const BLUE_WORDS = 9;
+export const RED_WORDS = 8;
+export const GRAY_WORDS = 7;
+export const BLACK_WORDS = 1;
+
+// Starting scores
+export const STARTING_SCORES: Scores = {
+  [Color.Blue]: BLUE_WORDS,
+  [Color.Red]: RED_WORDS,
+  [Color.Gray]: GRAY_WORDS,
+  [Color.Black]: BLACK_WORDS
+};
