+++ packages/frontend/src/index.ts
@@ -1,2 +1,4 @@
 export { Components, JSX } from "./components";
 export * from "./extra/types";
+export * from "./extra/utils";
+export * from "./extra/constants";
